Refactor: Move delegates to `InfiniFrame.Native.Delegates` namespace for improved organization
