+++ src/InfiniFrame.Shared/Native/Delegates/CppGetAllMonitorsDelegate.cs
@@ -3,7 +3,7 @@
 // ---------------------------------------------------------------------------------------------------------------------
 using System.Runtime.InteropServices;
 
-namespace InfiniFrame.Native;
+namespace InfiniFrame.Native.Delegates;
 // ---------------------------------------------------------------------------------------------------------------------
 // Code
 // ---------------------------------------------------------------------------------------------------------------------
